feat: Переосмыслен навигатор, добавлены команды
